+++ src/app/components/uiRPC.ts
@@ -1,21 +1,10 @@
-import { rpcStore } from '../models/rpc'
+import { TuneFormStore } from '../models/TuneFormStore'
+import { formProps } from '../../plugin/tune'
 
 export async function setAnimationTags(animationTag: string, delay: number, repeat: number) {
-  await rpcStore.setAnimationTags(animationTag, delay, repeat)
+  await TuneFormStore.setAnimationTags(animationTag, delay, repeat)
 }
 
-export async function updateProps(settings: {
-  shadowSize: number
-  brushSize: number
-  suggestedBrushSize: number
-  stepCount: number
-  stepNumber: number
-  displayMode: string
-  template: string
-  clearBefore: boolean
-  clearLayers: string[]
-  otherTags: string[]
-  brushType: string
-}) {
-  await rpcStore.updateProps(settings)
+export async function updateUiProps(settings: formProps) {
+  await TuneFormStore.updateProps(settings)
 }
